#213 fix(fe): 낙관적인 업데이트 함수 제거
